pyproject.toml
[project]
name = "disco"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py-self>=2.0.0",
    "python-dotenv>=1.0.0",
]
